♻️ refactor: petition 도메인 엔티티 리팩토링 (#3)
