--- src/main/java/com/smunity/petition/domain/Agreement.java
@@ -1,24 +1,0 @@
-package com.smunity.petition.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "petition_voter")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Agreement {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "agreement_id")
-    private Long agreementId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "petiion_id")
-    private Petition petition;
-
-}
