--- src/main/java/com/smunity/petition/domain/Comment.java
@@ -1,32 +1,0 @@
-package com.smunity.petition.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "comment")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Comment {
-
-    @Id @GeneratedValue
-    @Column(name = "comment_id")
-    private Long commentId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "author_id")
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "petiion_id")
-    private Petition petition;
-
-    @Column(name = "create_date")
-    private LocalDateTime createdDate;
-
-    @Column(name = "modified_date")
-    private LocalDateTime modifiedDate;
-
-}
