+++ src/main/java/com/smunity/petition/domain/petition/entity/Agreement.java
@@ -1,0 +1,28 @@
+package com.smunity.petition.domain.petition.entity;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Table(name = "petitions_petition_voter")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Agreement {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column
+    private Long user_id;
+
+    //TODO User 테이블 생성 후 연결
+//    @ManyToOne(fetch = FetchType.LAZY)
+//    @JoinColumn(name = "author_id")
+//    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "petition_id")
+    private Petition petition;
+}
