+++ src/main/java/com/smunity/petition/domain/petition/entity/Respond.java
@@ -1,23 +1,28 @@
-package com.smunity.petition.domain;
+package com.smunity.petition.domain.petition.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "answer")
+@Getter
+@Table(name = "petitions_answer")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Respond {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
 
-    @Id @GeneratedValue
-    @Column(name = "answer_id")
-    private Long answerId;
+    @Column
+    private Long author_id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "author_id")
-    private User user;
+    //TODO User 테이블 생성 후 연결
+//    @ManyToOne(fetch = FetchType.LAZY)
+//    @JoinColumn(name = "author_id")
+//    private User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "petition_id")
@@ -28,7 +33,6 @@ public class Respond {
     @Column(name = "create_date")
     private LocalDateTime createdDate;
 
-    @Column(name = "modified_date")
+    @Column(name = "modify_date")
     private LocalDateTime modifiedDate;
-
 }
